refactor(types): SkillIndex / Insight DTO 上移到 src/types/ 解开 renderer → server 反向依赖
renderer/skill-list-renderer / skill-detail-renderer 此前 import server/skill-index 和
server/skill-insights 拿稳定 DTO,违反「视图层不依赖装配层」边界。把 DTO 形状沉到
src/types/skill-index.ts,server 文件保留 runtime 逻辑且 re-export 同名类型保持向后兼容。

变更面:
- 新增 src/types/skill-index.ts,集中放 SkillDoctorSnapshot / SkillEvalSnapshot /
  SkillObserveSnapshot / SkillIndexEntry / SkillIndexSummary / SkillIndex 和 Insight*
  全部 11 个类型。
- StudioDiagnosisSummary 一并上移到 src/types/diagnosis.ts(SkillIndex 引用它),
  diagnosis/studio-projection.ts 改 re-export。
- types/report.ts 的旧 Insight 重命名为 AnalysisInsight,消除跟新 SkillInsight
  的 barrel 命名冲突。仅 src/analysis/report-diagnostics.ts 和 src/renderer/summary.ts
  两个内部消费者,JSON 字段语义不变。
- src/server/skill-index.ts 和 src/server/skill-insights.ts re-export 全部 DTO 供老
  路径使用,测试不受影响。
- renderer 两个文件 import 全部走 ../types/index.js,不再 reach 进 server/。

测量学守门:Report JSON schema 字段不变(只是 TS 接口改名),judge hash / observe
prompt hash / HTML snapshot 全部字节一致。

PR #159 的 boundary test 合入后,renderer/skill-*-renderer.ts::server/skill-* 的 4
条 whitelist entry 会被 dead-whitelist 检查自动标红,提示清理。
